ECG Application Draft
This is the application displaying the chart of the measured electrocardiogram (심전도 in Korean) This data is measured through ECG kit.
ECG kit is connected to the arduino UNO board. At the board, there is Bluetooth module, HC-06, which helps Bluetooth communication with other remote devices.
We transferred measured data through the module. (We have adapted UART communication)